% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{get_theme}
\alias{get_theme}
\title{Retrieve a list of thematic elements}
\usage{
get_theme(theme = "gray", base_size = NA, base_family = NA,
  horizontal = NA, dkpanel = NA, stata = NA, gray_bg = NA,
  bgcolor = NA, nomargin = NA, ff = NA, fc = NA, fs = NA, gM = NA,
  gm = NA, gc = NA, gl = NA, boxes = NA, bc = NA, pc = NA,
  lp = NA, axis = NA, light = NA, scheme = NA, ticks = NA,
  color = NA, title_family = NA, bg = NA, fg = NA, cex = NA,
  fill = NA, col = NA, lwd = NA, lty = NA)
}
\arguments{
\item{theme}{\code{character}. Theme from the \pkg{ggplot2} or \pkg{ggthemes}
packages. Defaults to 'gray'.}

\item{base_size}{base font size}

\item{base_family}{base font family}

\item{horizontal}{\code{logical}. Horizontal axis lines?}

\item{dkpanel}{\code{logical} Darker background for panel region?}

\item{stata}{\code{logical} Use RGB values from Stata's
economist scheme.}

\item{gray_bg}{\code{logical} If \code{TRUE}, use gray background, else use white
background.}

\item{bgcolor}{The background color of plot. One of \code{'default',
'darkunica'}, the names of values in
\code{ggthemes_data$hc$bg}.}

\item{nomargin}{suppress the white space around the plot (boolean)}

\item{ff}{font family, like \code{sans}. Deprecated: use \code{base_family} instead.}

\item{fc}{font color (name or hexa code)}

\item{fs}{font size (integer). Deprecated: use \code{base_size} instead.}

\item{gM}{major grid (boolean)}

\item{gm}{minor grid (boolean)}

\item{gc}{grid color (name or hexa code)}

\item{gl}{grid line type (\code{lty})}

\item{boxes}{to render a border around the plot or not}

\item{bc}{background color (name or hexa code)}

\item{pc}{panel background color (name or hexa code)}

\item{lp}{legend position}

\item{axis}{axis angle as defined in \code{par(les)}}

\item{light}{\code{logical}. Light or dark theme?}

\item{scheme}{One of "s2color", "s2mono", "s1color",
"s1rcolor", or "s1mono", "s2manual",
"s1manual", or "sj"}

\item{ticks}{\code{logical} Show axis ticks?}

\item{color}{The background color of plot. One of \code{'brown',
'gray', 'green', 'blue'}, the names of values in
\code{ggthemes_data$wsj$bg}.}

\item{title_family}{Plot title font family.}

\item{bg}{\code{character}. Background color for documentation.}

\item{fg}{\code{character}. Foreground/font color.}

\item{cex}{\code{numeric}. Font size multiplier.}

\item{fill}{\code{character}. Background color for plot.}

\item{col}{\code{character}. Border color.}

\item{lwd}{\code{numeric}. Border line width.}

\item{lty}{\code{character}, \code{numeric}. Border line type. Accepts 
numbers 0, 1, 2, 3, 4, 5, 6 or characters "blank", "solid", "dashed",
"dotted", "dotdash", "longdash", or "twodash".}
}
\description{
This function retrieves a list of thematic elements to be used by the
\link{doc_plot} function. The elements correspond to either a given theme
(such as from the \pkg{ggplot2} or \pkg{ggthemes} packages) or a custom theme
with relevant parameters also given.
}
