% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdocumentation.R
\docType{package}
\name{ggdocumentation}
\alias{ggdocumentation}
\alias{ggdocumentation-package}
\alias{ggdocumentation-package}
\title{Add documentation to ggplot2 plots}
\description{
Adds documentation elements such as sponsor, author, author job title, and
data sources to annotate \pkg{ggplot2} plots. Documentation is one of the 
principles of analytical design according to Edward Tufte's "Beautiful 
Evidence" - which inspired this package.
}
\details{
The only function you're likely to need from \pkg{ggdocumentation} is
\code{\link{doc_plot}}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/joshuaadampoirier/ggdocumentation}
}

}
\author{
\strong{Maintainer}: Joshua Poirier \email{josh.poirier@gmail.com}

}
