% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_plot.R
\name{doc_plot}
\alias{doc_plot}
\title{Add documentation to ggplot2 plots}
\usage{
doc_plot(g, author = "", author_title = "", data_source = "",
  date = FALSE, img_sponsor = "", sponsor = "", theme = "gray",
  draw = TRUE, ...)
}
\arguments{
\item{g}{A ggplot2 object to which documentation is to be added.}

\item{author}{A character object indicating the author of the plot.}

\item{author_title}{A character object indicating the job title of the author
of the plot.}

\item{data_source}{A character object indicating the sources of the data 
displayed in the plot.}

\item{date}{A logical indicating whether or not to add today's date as the 
publishing date of the plot. It defaults to FALSE.}

\item{img_sponsor}{A character object representing the filename/filepath to
a *.png image file representing the sponsor (ie. company logo)}

\item{sponsor}{A character object indicating the sponsor of the plot. Ignored
if img_sponsor is used.}

\item{theme}{A character object indicating the theme for the documentation.
This value should correspond to themes from the ggplot2 or ggthemes 
packages. It defaults to 'gray'.}

\item{draw}{A logical object indicating whether or not to draw the plot. If
FALSE the object to be plotted is returned without being drawn.}

\item{...}{Additional parameters to be passed into the get_theme() function.}
}
\value{
The object to be plotted.
}
\description{
Adds documentation elements such as sponsor, author, author job title, and
data sources to annotate ggplot2 plots. Documentation is one of the 
principles of analytical design according to Edward Tufte's "Beautiful 
Evidence".
}
\examples{
library("ggdocumentation")

png(filename="figures/iris_fivethirtyeight.png", 
    width=1100, height=800, units="px")
    
g <- qplot(Sepal.Length, Petal.Length, data=iris, color=Species)
d <- doc_plot(g)
plot_grid(d)
}
